requirements.txt
asgiref==3.6.0
asttokens==2.2.1
autopep8==2.0.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
gunicorn==20.1.0
ipykernel==6.21.0
ipython==8.9.0
jedi==0.18.2
jupyter_client==8.0.2
jupyter_core==5.2.0
matplotlib-inline==0.1.6
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
tzdata==2022.7
wcwidth==0.2.6
